Implement user registration page
Log in now required to add, edit, and delete rooms.
    - Logged in user must be the room host to edit or delete the room.

+++ main/views.py
@@ -1,15 +1,23 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
 
 
 def login_page(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -26,7 +34,7 @@ def login_page(request):
         else:
             messages.error(request, 'Username or password is incorrect')
 
-    context = {}
+    context = {'page': page}
 
     return render(request, 'main/login_register.html', context)
 
@@ -35,6 +43,26 @@ def logout_user(request):
     logout(request)
 
     return redirect('home')
+
+
+def register_user(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+
+            login(request, user)
+
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurring during registration.')
+
+    return render(request, 'main/login_register.html', {'form': form})
 
 
 def home(request):
@@ -59,6 +87,7 @@ def room(request, primary_key):
     return render(request, 'main/room.html', context)
 
 
+@login_required(login_url='/login')
 def create_room(request):
     form = RoomForm()
 
@@ -75,9 +104,13 @@ def create_room(request):
     return render(request, 'main/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def update_room(request, primary_key):
     room = Room.objects.get(id=primary_key)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to do that!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -92,8 +125,12 @@ def update_room(request, primary_key):
     return render(request, 'main/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def delete_room(request, primary_key):
     room = Room.objects.get(id=primary_key)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to do that!")
 
     if request.method == 'POST':
         room.delete()
